% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost-data.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult humans}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Adm.clerical}{something}
  \item{Bachelors}{person is a bachlor}
  \item{Male}{person is a bachlor}
  \item{Never.married}{person is a bachlor}
  \item{Not.in.family}{person is a bachlor}
  \item{State.gov}{person is a bachlor}
  \item{United.States}{person is a bachlor}
  \item{White}{person is a bachlor}
  \item{X..50K}{person is a bachlor}
  \item{X0}{person is a bachlor}
  \item{X13}{person is a bachlor}
  \item{X2174}{person is a bachlor}
  \item{X39}{person is a bachlor}
  \item{X40}{person is a bachlor}
  \item{X77516}{person is a bachlor}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
adult
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

