% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost-data.R
\docType{data}
\name{australian}
\alias{australian}
\title{Australian}
\format{An object of class \code{data.frame} with 689 rows and 15 columns.}
\usage{
australian
}
\description{
Australian
}
\keyword{datasets}

