% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost-data.R
\docType{data}
\name{coli2000}
\alias{coli2000}
\title{coli2000}
\format{An object of class \code{data.frame} with 9821 rows and 86 columns.}
\usage{
coli2000
}
\description{
coli2000
}
\keyword{datasets}

