% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\docType{class}
\name{Deepboost-class}
\alias{Deepboost-class}
\title{An S4 class to represent a bank account.}
\description{
An S4 class to represent a bank account.
}
\section{Slots}{

\describe{
\item{\code{lambda}}{parameter something}

\item{\code{train}}{deepboost model training function}

\item{\code{predict}}{deepboost model instance prediction function}

\item{\code{print}}{deepboost model evaluation statistics function}

\item{\code{error}}{deepboost model training error}
}}

