% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\name{deepboost.formula}
\alias{deepboost.formula}
\title{Main function for deepboost model creation, using a formula}
\usage{
deepboost.formula(formula, data, weights = NULL, tree_depth = 5,
  num_iter = 1, beta = 0, lambda = 0.05, loss_type = "l",
  verbose = TRUE)
}
\arguments{
\item{formula}{A R Formula object see : ?formula}

\item{data}{A data.frame of samples to train on}

\item{weights}{The weight of each example}

\item{controls}{parameters}
}
\value{
A trained Deepbost model
}
\description{
Main function for deepboost model creation, using a formula
}

