% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\name{deepboost.predict}
\alias{deepboost.predict}
\title{Predicts instances responses based on a deepboost model}
\usage{
deepboost.predict(object, newdata)
}
\arguments{
\item{object}{A Deepboost S4 class object}

\item{newdata}{A data.frame to predict responses for}
}
\value{
A vector of respones
}
\description{
Predicts instances responses based on a deepboost model
}

