% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\name{deepboost.print}
\alias{deepboost.print}
\title{Evaluates and prints statistics for a deepboost model on the train set}
\usage{
deepboost.print(object)
}
\arguments{
\item{object}{A Deepboost S4 class object}
}
\value{
List with model_statistics to console the model evaluation string
}
\description{
Evaluates and prints statistics for a deepboost model on the train set
}

