% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\docType{data}
\name{Deepboost}
\alias{Deepboost}
\title{Empty Deepboost S4 class object with default settings}
\format{An object of class \code{Deepboost} of length 1.}
\usage{
Deepboost
}
\description{
Empty Deepboost S4 class object with default settings
}
\keyword{datasets}

