% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\name{deepboost.train}
\alias{deepboost.train}
\title{Trains a deepboost model}
\usage{
deepboost.train(object, data, tree_depth, num_iter, beta, lambda, loss_type,
  verbose, classes)
}
\arguments{
\item{object}{A Deepboost S4 class object}
}
\value{
A trained Deepbost model
}
\description{
Trains a deepboost model
}
\details{
(beta,lambda) = (0,0) - adaboost, (>0,0) - L1, (0,>0) deepboost, (>0, >0) deepbost+L1
}

