% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\docType{methods}
\name{show,Deepboost-method}
\alias{show,Deepboost-method}
\title{Print method for Deepboost model}
\usage{
\S4method{show}{Deepboost}(object)
}
\description{
Evaluates a trained deepboost model object.
}
\details{
Evalutes a Deepboost model object.
Prints : bla \code{error}
}

