% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\docType{class}
\name{Deepboost-class}
\alias{Deepboost-class}
\title{An S4 class to represent a deepboost model.}
\description{
An S4 class to represent a deepboost model.
}
\section{Slots}{

\describe{
\item{\code{tree_depth}}{maximum depth for a single decision tree in the model}

\item{\code{num_iter}}{number of iterations = number of trees in ensemble}

\item{\code{beta}}{regularisation for scores (L1)}

\item{\code{lambda}}{regularisation for tree depth}

\item{\code{loss_type}}{"l" logistic, "e" exponential}

\item{\code{verbose}}{print extra data while training TRUE / FALSE}

\item{\code{examples}}{data.frame with instances used for model training}

\item{\code{model}}{Deepboost model as used by C code serialised to R List}

\item{\code{classes}}{a vector of factors representing the classes used for classification with this model}
}}

