% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\name{deepboost.default}
\alias{deepboost.default}
\title{Main function for deepboost model creation}
\usage{
deepboost.default(x, y, weights = NULL, controls = NULL)
}
\arguments{
\item{x}{A data.frame of samples' values}

\item{y}{A data.frame of samples's labels}

\item{weights}{The weight of each example}

\item{controls}{parameters}
}
\value{
A trained Deepbost model
}
\description{
Main function for deepboost model creation
}

