% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\name{deepboost}
\alias{deepboost}
\title{Main function for deepboost model creation}
\usage{
deepboost(formula, data, ...)
}
\arguments{
\item{formula}{A R Formula object see : ?formula}

\item{data}{A data.frame of samples to train on}

\item{weights}{The weight of each example}

\item{controls}{parameters}
}
\value{
A trained Deepbost model
}
\description{
Main function for deepboost model creation
}

