% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\name{deepboost.train}
\alias{deepboost.train}
\title{Trains a deepboost model}
\usage{
deepboost.train(object, data, controls, weights = NULL, fitmem = NULL, ...)
}
\arguments{
\item{object}{A Deepboost S4 class object}

\item{data}{A data.frame to train on}

\item{controls}{Paramters}
}
\value{
A trained Deepbost model
}
\description{
Trains a deepboost model
}

