% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepboost.R
\docType{methods}
\name{show,Deepboost-method}
\alias{show,Deepboost-method}
\title{Print method for Deepboost model
Evaluates a trained deepboost model object.}
\usage{
\S4method{show}{Deepboost}(object)
}
\arguments{
\item{object}{Object of class "Deepboost"}
}
\description{
Print method for Deepboost model
Evaluates a trained deepboost model object.
}
\details{
Prints :
Model error: X"
Average tree size: Y"
Number of trees: Z"
}

